requirements.txt
fastapi
uvicorn
langserve
langchain
langchain-community
langchain-core
langchain-openai
openai
tiktoken
pydantic
chromadb
sse-starlette
python-dotenv